ui: Fix deviceScale for mobiles

+++ internal/ui/ui_android.go
@@ -14,6 +14,6 @@
 
 package ui
 
-func deviceScale() int {
+func deviceScale() float64 {
 	return 1
 }
